requirements.txt
coverage==7.6.1
discord.py==2.4.0
dpytest==0.7.0
flake8==7.1.1
pytest==8.3.2
pytest-asyncio==0.23.8
pytest-cov==5.0.0
schema==0.7.7